update imports to reflect refactored helper modules
